pyproject.toml
[tool.poetry]
name = "criticalpath_ai"
version = "0.1.0"
description = "Criticalpath AI with Graph KnowledgeBase"
readme = "README.md"
authors = ["Priyank"]

packages = [
  { include = "backend" },
#  { include = "" }
]
[tool.poetry.dependencies]
python = ">=3.11,<3.15"
anthropic = "^0.75.0"
google-genai = "^1.56.0"
openai = "^2.14.0"
portkey-ai = "^2.1.0"
python-dotenv = "^1.2.1"
fastapi = {extras = ["standard"], version = "^0.128.7"}
uvicorn = { version = "^0.27", extras = ["standard"] }
pandas = "^2.2"
numpy = "^1.26"
networkx = "^3.2"
pydantic = "^2.6"
pytest = "^9.0.2"

# Optional deps (to power extras like .[modal])
modal = { version = "^0.73.0", optional = true }
prime-sandboxes = { version = "^0.2.0", optional = true }
dill = { version = "^0.3.7", optional = true }

[tool.poetry.extras]
modal = ["modal", "dill"]
prime = ["prime-sandboxes", "dill"]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.1"
ruff = "^0.14.10"
ty = "^0.0.7"

[tool.poetry.group.test.dependencies]
pytest-asyncio = "^1.3.0"
pytest = "^9.0.2"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]


#[tool.poetry.scripts]
#criticalpath = "criticalpath_ai.cli:main"